# packages/swagger-convert-core/test/swagger/index.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Your API Title",
    "description": "Your API Description",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
  "tags": [
    {
      "name": "get-request",
      "description": "Get Request"
    }
  ],
  "paths": {
    "/getRequest/getBody": {
      "get": {
        "tags": [
          "get-request"
        ],
        "summary": "bodyListItems",
        "operationId": "bodyListItemsUsingGET",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "length",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemBody"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/getRequest/search": {
      "get": {
        "tags": [
          "get-request"
        ],
        "summary": "searchItems",
        "operationId": "searchItemsUsingGET",
        "parameters": [
          {
            "name": "length",
            "in": "query",
            "description": "length",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "query",
            "required": true,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "text",
            "in": "query",
            "description": "text",
            "required": true,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/getRequest/users/{id}": {
      "get": {
        "tags": [
          "get-request"
        ],
        "summary": "getUserById",
        "operationId": "getUserByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/getRequest/users/{id}/{spanceId}": {
      "get": {
        "tags": [
          "get-request"
        ],
        "summary": "getUserByIds",
        "operationId": "getUserByIdsUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "spanceId",
            "in": "path",
            "description": "spanceId",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Item": {
        "title": "Item",
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "ItemBody": {
        "title": "ItemBody",
        "type": "object",
        "properties": {
          "listItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "name": {
            "type": "string",
            "description": "名称"
          },
          "text": {
            "type": "string",
            "description": "描述"
          }
        }
      }
    }
  }
}